archive/prototype-rest/services/iris/requirements.txt
fastapi==0.116.1
uvicorn==0.35.0
open-iris==1.10.0
numpy==2.2.6
pillow==11.3.0
